Move db to lib directory
